requirements.txt
boto3==1.17.44
PyJWT==2.0.1
cryptography==3.4.7
six==1.15.0
requests==2.25.1
